Account view and edit are in one page now
